Finishing up on SubSpace.

+++ src/world/mod.rs
@@ -10,6 +10,6 @@ pub use self::tile::{TileID, TileName};
 /// So, the length along each axis of our chunk is then 2^(CHUNK_SCALE) meters.
 /// This also corresponds to the height of the root node on an octree: 
 /// CHUNK_SCALE steps down from the root node is your 1x1x1 meter cube.
-pub const CHUNK_SCALE : usize = 6;
+pub const CHUNK_SCALE : i8 = 6;
 //Good candidates are 5 (32 meters to a side), 6 (64 meters to a side),
 //or 8 (256 meters to a side).
